requirements.txt
fprs==1.0.0
moojoco==1.1.6
numpy==2.1.2
transforms3d==0.4.2
mujoco==3.3.0
opencv-python==4.10.0.84
pywavefront==1.3.3
pillow==10.4.0
trimesh==4.4.9
cmaes==0.11.1
scipy==1.14.1
tqdm==4.66.5
wandb==0.19.6
matplotlib==3.10.0
seaborn==0.13.2
fpdf==1.7.2
pandas==2.2.3